move CipherBetweenTwo from datastructure to etc
